+++ src/main/java/es/in2/issuer/backend/shared/domain/exception/JWTVerificationException.java
@@ -1,8 +1,7 @@
 package es.in2.issuer.backend.shared.domain.exception;
 
-import org.springframework.security.core.AuthenticationException;
 
-public class JWTVerificationException extends AuthenticationException {
+public class JWTVerificationException extends RuntimeException {
     public JWTVerificationException(String message) {
         super(message);
     }